requirements.txt
python-dotenv==1.0.1
pyyaml==6.0.1
jsonschema==4.23.0
rich==13.8.1
snowflake-connector-python==3.12.3
streamlit==1.21.0
pandas==2.2.3
networkx==3.3
